+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Improvements and new features:
 - Solve problems with file name limitations and document the used options of ``gcov``. (:issue:`528`)
 - Add :option:`--no-markers` to ignore exclusion markers in code. (:issue:`361`)
 - Generate also a Report of Excluded Coverage. (:issue:`503`)
+- Added :option:`--decisions` to add decision coverage to HTML and JSON output. (:issue:`350`)
 
 Documentation: